requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bleach==4.1.0
blinker==1.4
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
cleo==0.8.1
click==6.7
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
crashtest==0.3.1
cryptography==36.0.1
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.12
django-cors-headers==3.11.0
django-extensions==2.2.9
django-filter==21.1
django-nose==1.4.6
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
drf-localize==0.1.5
drf-util==0.0.12
drf-yasg==1.20.0
easy-colors==0.0.1
elasticsearch==6.3.1
entrypoints==0.3
Faker==4.1.0
filelock==3.4.0
flex==6.13.2
html5lib==1.1
idna==2.7
importlib-metadata==4.11.1
inflection==0.5.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonpointer==1.14
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
keyring==23.5.0
lockfile==0.12.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
oauthlib==3.2.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pastel==0.2.1
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.2
platformdirs==2.4.0
poetry==1.1.13
poetry-core==1.0.8
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pylev==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==3.13
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
requests==2.20.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3987==1.3.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
send-mail==0.0.2
Send2Trash==1.8.0
shellingham==1.4.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stevedore==3.5.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.4
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tomlkit==0.10.0
tornado==6.1
traitlets==5.1.0
uritemplate==4.1.1
urllib3==1.24.3
validate-email==1.3
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1

